+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the "mme2k-powerapps-helper" extension will be documented in this file.
 
+## 0.7.0
+
+- PowerApps Helper reference now [Power Platform Extension](https://github.com/microsoft/powerplatform-vscode) from Microsoft and use the included/installed PAC CLI. It is installed e.g. in Windows underneath `c:\Users\<YOUR USERNAME>\AppData\Roaming\<VS CODE FOLDER>\User\globalStorage\microsoft-isvexptools.powerplatform-vscode\pac\tools\`.
+- Usage of 'keytar' wrapped into try-catch to support MacOS
+
 ## 0.6.5
 
 - Rebuild for Power Platform Cli (1.9.4)